requirements.txt
flask>=3.0.0
langgraph>=0.2.0
langchain>=0.2.0,<0.3.0
langchain-core>=0.2.0
langchain-openai>=0.1.0
langchain-community>=0.2.0
openai>=1.12.0
python-dotenv>=1.0.0
werkzeug>=3.0.0
pandas>=2.0.0
numpy>=1.24.0

